requirements.txt
asgiref==3.2.10
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.1
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-session-timeout==0.1.0
idna==3.3
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
gunicorn==20.1.0 
